requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
lxml==5.3.0
slowapi==0.1.9
redis==5.0.1
sentry-sdk==1.50.0
gunicorn==23.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

